refactor(cli-core): move exports from index.ts to _exports/index.ts
Moves all exports to _exports/index.ts and removes the old index.ts,
aligning with the existing _exports/ convention used by ux, request,
and package-manager sub-exports. Updates tsconfig path aliases
accordingly.
