+++ packages/@sanity/cli-core/package.json
@@ -25,12 +25,12 @@
   ],
   "type": "module",
   "sideEffects": false,
-  "main": "./dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "./dist/_exports/index.js",
+  "types": "dist/_exports/index.d.ts",
   "exports": {
     ".": {
-      "source": "./src/index.ts",
-      "default": "./dist/index.js"
+      "source": "./src/_exports/index.ts",
+      "default": "./dist/_exports/index.js"
     },
     "./ux": {
       "source": "./src/_exports/ux.ts",
