+++ packages/@sanity/cli/tsconfig.json
@@ -6,7 +6,7 @@
     "rootDir": ".",
     "jsx": "react-jsx",
     "paths": {
-      "@sanity/cli-core": ["./node_modules/@sanity/cli-core/src/index.ts"],
+      "@sanity/cli-core": ["./node_modules/@sanity/cli-core/src/_exports/index.ts"],
       "@sanity/cli-core/*": ["./node_modules/@sanity/cli-core/src/_exports/*"],
       "@sanity/cli-test": ["./node_modules/@sanity/cli-test/src/index.ts"],
       "@sanity/cli-test/vitest": ["./node_modules/@sanity/cli-test/src/vitest.ts"]
